Move newly introduced logging condition and Log4j2, Logback classes to the hibernate package
